backend/requirements/dev.txt
-r base.txt
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
pytest-cov>=5.0
factory-boy>=3.3
faker>=25.0
django-debug-toolbar>=4.3
ipdb>=0.13
ruff>=0.9
mypy>=1.15
django-stubs>=5.1
djangorestframework-stubs>=3.15
django-filter-stubs>=0.1.3
pre-commit>=4.0
